Initial Commit
Editor and code base to allow for quick and easy creation of choose your own adventure games. Ideally used with IDLE. If you encounter issues with the pauses, just remove the time.sleep(0.04) on line 44, in Dialogue.Type(), dialogue.py (this is used to create the typing effect and add pauses, which can make things more readable.)
README coming soon.